fix: remove uint cast - CsWin32 expects enum, not uint

+++ LibreHardwareMonitorLib/Interop/PowerMonitor/Stm32PortFinder.cs
@@ -112,7 +112,7 @@ public static class Stm32PortFinder
                 if (PInvoke.SetupDiGetDeviceRegistryProperty(
                         (Windows.Win32.Devices.DeviceAndDriverInstallation.HDEVINFO)hDevInfo.DangerousGetHandle(),
                         &devInfoData,
-                        (uint)property,
+                        property,
                         null,
                         pBuffer,
                         BufferSize,
